erp-product-request/requirements.txt
pymongo
python-dotenv
fastapi
uvicorn